requirements.txt
Django>=5.2,<6.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-yasg>=1.21,<2.0
psycopg[binary]>=3.2,<4.0
Pillow>=10.0,<12.0

